# webapp/src/main/webapp/docs/swagger.yaml
openapi: 3.0.1
info:
  title: HealthSync API
  version: 1.0.0
...
paths:
...
  /doctors:
    get:
      operationId: listDoctors
      description: Returns a filtered page of doctors
      tags:
        - doctors
      parameters:
        - name: page
          in: query
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 1
        - name: pageSize
          in: query
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 10
        - name: name
          in: query
          schema:
            type: string
            example: John Doe
        - name: date
          in: query
          schema:
            format: date
            type: string
            example: 2024-07-25
        - name: fromTime
          in: query
          schema:
            allOf:
              - $ref : '#/components/schemas/TimeBlock'
        - name: toTime
          in: query
          schema:
            allOf:
              - $ref : '#/components/schemas/TimeBlock'
        - name: specialty
          in: query
          schema:
            uniqueItems: true
            type: array
            items:
              allOf:
                - $ref : '#/components/schemas/Specialty'
        - name: city
          in: query
          schema:
            uniqueItems: true
            type: array
            items:
              type: string
        - name: healthInsurance
          in: query
          schema:
            uniqueItems: true
            type: array
            items:
              allOf:
                - $ref : '#/components/schemas/HealthInsurance'
        - name: minRating
          in: query
          schema:
            minimum: 1
            maximum: 5
            type: integer
            format: int32
      responses:
        '200':
          description: Returned page of doctors
          content:
            application/vnd.doctor-list.v1+json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref : '#/components/schemas/Doctor'
        '204':
          description: No content for specified page and filters
    post:
      operationId: createDoctor
      description: Creates a doctor
      tags:
        - doctors
      requestBody:
        content:
          application/vnd.doctor.v1+json:
            schema:
              $ref: '#/components/schemas/DoctorRegisterForm'
      responses:
        '201':
          description: Doctor Created
        '400':
          description: Invalid request
          content:
            application/vnd.validation-error-list.v1+json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/ValidationError'
        '409':
          description: Email already in use
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
  /doctors/{doctorId}:
    get:
      operationId: getDoctor
      description: Gets the doctor
      tags:
        - doctors
      parameters:
        - name: doctorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Returns the doctor
          content:
            application/vnd.doctor.v1+json:
              schema:
                allOf:
                  - $ref : '#/components/schemas/Doctor'
        '404':
          description: Doctor not found
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
    put:
      operationId: updateDoctor
      description: Updates the doctor
      tags:
        - doctors
      parameters:
        - name: doctorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          multipart/form-data:
            schema:
              allOf:
                - $ref : '#/components/schemas/DoctorEditForm'
      responses:
        '204':
          description: Doctor updated successfully
        '400':
          description: Invalid request
          content:
            application/vnd.validation-error-list.v1+json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/ValidationError'
        '403':
          description: Only the doctor can update itself
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
        '404':
          description: Doctor not found
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
        '409':
          description: Email already in use
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
...
  /doctors/{doctorId}/reviews:
    get:
      operationId: listReviews
      description: Paginated list of reviews for the doctor
      tags:
        - reviews
      parameters:
        - name: doctorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 1
        - name: pageSize
          in: query
          schema:
            minimum: 1
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          description: Review page
          content:
            application/vnd.reviews-list.v1+json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref : '#/components/schemas/Review'
        '204':
          description: No content for page
        '404':
          description: Review or doctor not found
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
    post:
      operationId: createReview
      description: Creates a review for the doctor
      tags:
        - reviews
      parameters:
        - name: doctorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/vnd.review.v1+json:
            schema:
              $ref: '#/components/schemas/ReviewForm'
      responses:
        '201':
          description: Review Created
        '400':
          description: Invalid request
          content:
            application/vnd.validation-error-list.v1+json:
              schema:
                type: array
                items:
                  allOf:
                    - $ref: '#/components/schemas/ValidationError'
        '403':
          description: Only a user that had an appointment with the doctor can review
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
        '404':
          description: Doctor not found
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
  /doctors/{doctorId}/reviews/{reviewId}:
    get:
      operationId: getReview
      description: Gets the review
      tags:
        - reviews
      parameters:
        - name: doctorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: reviewId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Review
          content:
            application/vnd.review.v1+json:
              schema:
                allOf:
                  - $ref : '#/components/schemas/Review'
        '404':
          description: Review or doctor not found
          content:
            application/vnd.error.v1+json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    DoctorRegisterForm:
      required:
        - address
        - city
        - confirmPassword
        - email
        - healthInsurances
        - lastname
        - name
        - password
        - specialty
      type: object
      properties:
        password:
          maxLength: 50
          minLength: 4
          pattern: '[a-zA-Z0-9]+'
          type: string
          example: 'johndoesecret1234'
        confirmPassword:
          maxLength: 50
          minLength: 4
          pattern: '[a-zA-Z0-9]+'
          type: string
          example: 'johndoesecret1234'
        name:
          maxLength: 50
          minLength: 1
          pattern: '[a-zA-Z ñÑáÁéÉíÍóÓúÚ]+'
          type: string
          example: 'john'
        lastname:
          maxLength: 50
          minLength: 1
          pattern: '[a-zA-Z ñÑáÁéÉíÍóÓúÚ]+'
          type: string
          example: 'doe'
        email:
          maxLength: 50
          minLength: 1
          pattern: '[a-zA-Z0-9.+-ñÑ]+@[a-zA-Z0-9.-]+(.com|.com.ar|.edu.ar)'
          type: string
          example: 'john@gmail.con'
        healthInsurance:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/HealthInsurance'
          example: ['OSDE', 'OMINT']
        city:
          maxLength: 20
          minLength: 1
          pattern: '[a-zA-Z0-9. ñÑáÁéÉíÍóÓúÚ]+'
          type: string
          example: 'CABA'
        address:
          maxLength: 100
          minLength: 1
          pattern: '[a-zA-Z0-9. ñÑáÁéÉíÍóÓúÚ]+'
          type: string
          example: 'San Martin 202'
        specialty:
          allOf:
            - $ref: '#/components/schemas/Specialty'
    DoctorEditForm:
      required:
        - email
        - healthInsurance
        - lastname
        - locale
        - name
        - specialty
        - address
        - city
      type: object
      properties:
        image:
          type: string
          format: binary
        name:
          maxLength: 50
          minLength: 1
          pattern: '[a-zA-Z ñÑáÁéÉíÍóÓúÚ]+'
          type: string
          example: 'juan'
        lastname:
          maxLength: 50
          minLength: 1
          pattern: '[a-zA-Z ñÑáÁéÉíÍóÓúÚ]+'
          type: string
          example: 'doe'
        email:
          maxLength: 50
          minLength: 1
          pattern: '[a-zA-Z0-9.+-ñÑ]+@[a-zA-Z0-9.-]+(.com|.com.ar|.edu.ar)'
          type: string
          example: 'juan@gmail.com'
        locale:
          type: string
          example: 'es'
        healthInsurance:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/HealthInsurance'
          example: ['OSDE', 'OMINT']
        city:
          maxLength: 20
          minLength: 1
          pattern: '[a-zA-Z0-9. ñÑáÁéÉíÍóÓúÚ]+'
          type: string
          example: 'CABA'
        address:
          maxLength: 100
          minLength: 1
          pattern: '[a-zA-Z0-9. ñÑáÁéÉíÍóÓúÚ]+'
          type: string
          example: 'San Martin 202'
        specialty:
          allOf:
            - $ref: '#/components/schemas/Specialty'
...
    ReviewForm:
      required:
        - description
        - rating
      type: object
      properties:
        description:
          example: "Gran doctor, le doy 5 estrellas"
          maxLength: 1000
          minLength: 1
          type: string
        rating:
          maximum: 5
          minimum: 1
          type: integer
          format: int32
...
    Error:
      type: object
      properties:
        message:
          type: string
    ValidationError:
      type: object
      properties:
        attribute:
          type: string
        invalidValue:
          type: string
        message:
          type: string
...
    Review:
      type: object
      properties:
        date:
          type: string
          format: date
          example: "2023-05-17"
        description:
          type: string
          example: "Genial! El doctor me arreglo el pie."
        doctor:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/doctors/20"
        id:
          type: integer
          format: int64
          example: 1
        patient:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/patients/21"
        rating:
          type: integer
          format: int32
          example: 5
        self:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/doctors/20/reviews/1"
...
    Doctor:
      type: object
      properties:
        address:
          type: string
          example: "San Martin 202"
        appointments:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/appointments?userId=20"
        attendingHours:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/doctors/20/attendinghours"
        occupiedHours:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/doctors/20/occupiedhours"
        city:
          type: string
          example: "CABA"
        email:
          type: string
          format: email
          example: "john@gmail.com"
        firstName:
          type: string
          example: "john"
        healthInsurances:
          type: array
          items:
            type: string
            format: uri
            example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/healthinsurances/3"
          example:
            - "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/healthinsurances/3"
            - "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/healthinsurances/2"
            - "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/healthinsurances/1"
            - "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/healthinsurances/0"
        lastName:
          type: string
          example: "doe"
        notifications:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/notifications?userId=20"
        rating:
          type: number
          format: float
          example: 3.625
        ratingCount:
          type: integer
          format: int32
          example: 8
        reviews:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/doctors/20/reviews"
        self:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/doctors/20"
        specialty:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/specialties/17"
        vacations:
          type: string
          format: uri
          example: "http://old-pawserver.it.itba.edu.ar/paw-2023a-02/api/doctors/20/vacations"
...
    TimeBlock:
      type: string
      enum:
        - "00:00"
        - "00:30"
        - "01:00"
        - "01:30"
        - "02:00"
        - "02:30"
        - "03:00"
        - "03:30"
        - "04:00"
        - "04:30"
        - "05:00"
        - "05:30"
        - "06:00"
        - "06:30"
        - "07:00"
        - "07:30"
        - "08:00"
        - "08:30"
        - "09:00"
        - "09:30"
        - "10:00"
        - "10:30"
        - "11:00"
        - "11:30"
        - "12:00"
        - "12:30"
        - "13:00"
        - "13:30"
        - "14:00"
        - "14:30"
        - "15:00"
        - "15:30"
        - "16:00"
        - "16:30"
        - "17:00"
        - "17:30"
        - "18:00"
        - "18:30"
        - "19:00"
        - "19:30"
        - "20:00"
        - "20:30"
        - "21:00"
        - "21:30"
        - "22:00"
        - "22:30"
        - "23:00"
        - "23:30"
    HealthInsurance:
      type: string
      enum:
        - NONE
        - OMINT
        - OSDE
        - SWISS_MEDICAL
        - GALENO
        - MEDIFE
        - MEDICUS
        - ACCORD
        - HOSPITAL_ITALIANO
        - HOSPITAL_ALEMAN
        - HOSPITAL_BRITANICO
        - PREVENCION_SALUD
        - AVALIAN
        - WILLIAM_HOPE
        - PRE_MEDIC
        - OSFE
        - OSPACA
        - ANDAR
        - LUIS_PASTEUR
        - UNION_PERSONAL
        - JERARQUICOS
        - OSDEPYM
        - OSPE
        - OSPLAD
        - OSSEG
        - SANCOR_SALUD
    Specialty:
      type: string
      enum:
        - ALLERGY_AND_IMMUNOLOGY
        - PEDIATRIC_ALLERGY_AND_IMMUNOLOGY
        - PATHOLOGICAL_ANATOMY
        - ANESTHESIOLOGY
        - GENERAL_ANGIOLOGY_AND_HEMODYNAMICS
        - CARDIOLOGY
        - CHILDRENS_CARDIOLOGIST
        - CARDIOVASCULAR_SURGERY
        - PEDIATRIC_CARDIOVASCULAR_SURGERY
        - HEAD_AND_NECK_SURGERY
        - THORAX_SURGERY
        - GENERAL_SURGERY
        - CHILD_SURGERY
        - PLASTIC_AND_RECONSTRUCTIVE_SURGERY
        - PERIPHERAL_VASCULAR_SURGERY
        - MEDICAL_CLINIC
        - COLOPROCTOLOGY
        - DERMATOLOGY
        - PEDIATRIC_DERMATOLOGY
        - IMAGE_DIAGNOSIS
        - CARDIAC_ELECTROPHYSIOLOGY
        - EMERGENTOLOGY
        - ENDOCRINOLOGY
        - INFANT_ENDOCRINOLOGIST
        - CLINICAL_PHARMACOLOGY
        - PHYSIATRY
        - GASTROENTEROLOGY
        - CHILDRENS_GASTROENTEROLOGIST
        - MEDICAL_GENETICS
        - GERIATRICS
        - GYNECOLOGY
        - PEDIATRIC_HEMATO_AWARENESS
        - HEMATOLOGY
        - HEMOTHERAPY_AND_IMMUNOHEMATOLOGY
        - HEPATOLOGY
        - PEDIATRIC_HEPATOLOGY
        - INFECTOLOGY
        - CHILD_INFECTOLOGIST
        - SPORTS_MEDICINE
        - WORK_MEDICINE
        - GENERAL_MEDICINE_AND_OR_FAMILY_MEDICINE
        - LEGAL_MEDICINE
        - NUCLEAR_MEDICINE
        - PALITATIVE_MEDICINE
        - NEPHROLOGY
        - CHILD_NEPHROLOGIST
        - NEONATOLOGY
        - PNEUMONOLOGY
        - CHILD_PNEUMONOLOGIST
        - NEUROSURGERY
        - NEUROLOGY
        - CHILD_NEUROLOGIST
        - NUTRITION
        - OBSTETRICS
        - OPHTHALMOLOGY
        - ONCOLOGY
        - ORTHOPEDICS_AND_TRAUMATOLOGY
        - ORTHOPEDICS_AND_CHILD_TRAUMATOLOGY
        - OTORHINOLARYNGOLOGY
        - PEDIATRICS
        - PSYCHIATRY
        - CHILD_AND_YOUTH_PSYCHIATRY
        - RADIOTHERAPY_OR_RADIATION_THERAPY
        - RHEUMATOLOGY
        - CHILD_RHEUMATOLOGIST
        - INTENSIVE_THERAPY
        - INTENSIVE_CHILDRENS_THERAPIST
        - TOCOGINECOLOGY
        - TOXICOLOGY
        - UROLOGY